+++ doc/source/CHANGELOG.rst
@@ -1,6 +1,20 @@
 Changelog
 =========
 
+1.2.4 (11-12-18)
+---------------
+
+**Fixes:**
+- Effective count matrix parallel evaluation fixes. #116, #117
+
+1.2.3 (7-27-18)
+---------------
+
+**New features**
+
+- Effective count matrix can be evaluated in parallel. #112
+
+
 1.2.2 (6-25-18)
 ---------------